A lighter code to poke recruiters
This code just does the job of poking all the recruiters in an area who match a certain criteria. The current URl is set to work for recruiters in NYC. Please change as necessary. Yes the while loop will just run until linkedin runs out of recruiters in NYC and then it will throw an error.